requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
idna==2.8
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
Jinja2==2.10.1
json5==0.8.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.5.0
jupyterlab==1.0.1
jupyterlab-server==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
pyrsistent==0.15.2
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
pyzmq==18.0.2
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
